feat: update parser grammar and related files
