--- frontend/src/query-grammar/.antlr/FilterQueryLexer.tokens
@@ -1,49 +1,0 @@
-LPAREN=1
-RPAREN=2
-LBRACK=3
-RBRACK=4
-COMMA=5
-EQUALS=6
-NOT_EQUALS=7
-NEQ=8
-LT=9
-LE=10
-GT=11
-GE=12
-LIKE=13
-NOT_LIKE=14
-ILIKE=15
-NOT_ILIKE=16
-BETWEEN=17
-NOT_BETWEEN=18
-EXISTS=19
-NOT_EXISTS=20
-REGEXP=21
-NOT_REGEXP=22
-CONTAINS=23
-NOT_CONTAINS=24
-IN=25
-NOT_IN=26
-NOT=27
-AND=28
-OR=29
-HAS=30
-HASANY=31
-HASALL=32
-HASNONE=33
-BOOL=34
-NUMBER=35
-QUOTED_TEXT=36
-KEY=37
-WS=38
-'('=1
-')'=2
-'['=3
-']'=4
-','=5
-'!='=7
-'<>'=8
-'<'=9
-'<='=10
-'>'=11
-'>='=12
